rework tenant status
The status section of Tenant custom resources currently uses a state
machine enumeration field (`status.progress`). The [Kubernetes API
Conventions][1] discourage the use of state machine fields in favor of
using Conditions.

Currently the Tenant resource is implemented as a batch resource
(processed to completion once), but actually it is a living resource
where the controller should permanently respond to deviation of the
actual state from the desired state:

- Update the role binding of a tenant namespace if needed, e.g. if an
  operator has configured another tenant role or the role binding
  has been accidently deleted.

- Signal the situation that an assigned tenant namespace does not exist
  anymore in the status section of tenant resources.

To address the issues described above, this change:

- introduces a Ready condition in the status section of Tenant
  resources, replacing the status fields `progress`, `message` and
  `result`.

- implements a true reconciliation logic in the tenant controller.

[1]: https://github.com/kubernetes/community/blob/a2cdce51a0bbbc214f0e8813e0a877176ad3b6c9/contributors/devel/sig-architecture/api-conventions.md#typical-status-properties