runctl: run JFR container with service account 'default'
Currently a service account `run-bot` is used for the Jenkinsfile Runner
pod, while the `default` service account of the run namespace has no
permissions.

Pipeline implementations might need to start additional pods that should
have the same cluster API access as the JFR pod. To do so, they must
specify `run-bot` as the pod's `spec.serviceAccountName`. Omitting the
field results in running the additional pod as the `default` service
account, which has no cluster API permissions. This is considered as a
common mistake and should be avoided.

If we use the `default` service account instead of `run-bot` (having the
same permissions assigned as `run-bot` had; run JFR pod as `default`
service account), pipeline implementations don't have to deal with the
service account name when starting additional pods.