handle pipeline run details as optional
The run controller handles all fields of PipelineRunDetails as if they
are set, although all fields are optional and should not be passed as
Tekton TaskRun parameters if not set. Check for empty/unset values and
pass them to the Tekton TaskRun only if set.

Defining `sequenceNumber` as `int` is not appropriate for an API,
because it can be either 32-bit or 64-bit depending on the compilation
target architecture. For an API the size should be fixed. Use `int32`
instead.