simplify secret transformer functions
Instead of defining a rule that transformers must not modify the passed
object and return a pointer to a possibly modified copy, it's much
simpler, efficient and less error prone to let the function just modify
the object passed in. Then the caller is responsible for creating a copy
if needed.