add pipeline run timeout
So far we relied on Tekton's default timeout implicitly used for Steward
pipeline runs.

As the use of Tekton is an implementation detail, we should not require
user to configure Tekton in a Steward-specific way.

Add a pipeline run configuration option and a corresponding Helm chart
value to set the timeout for Steward pipeline runs.

Co-authored-by: Andreas Brehmer <andreas.brehmer@sap.com>