polish pkg/runctl/cfg
- do not exclude the default network policy from
  `PipelineRunsConfigStruct.NetworkPolicies`

- do not trim network policy values (could invalidate correct original
  values, e.g. if all lines are indented)

- `LoadPipelineRunsConfig`:

    - fail if the network policies config map does not exist (having no
      network policy would forbid any network traffic for pipeline
      run pods, which in turn would always fail)

- more precise error messages

- a lot of refactoring

PipelineRunsConfigStruct: default network profile: store key instead of value

In `PipelineRunsConfigStruct` the default network policy
is stored twice:

- as entry in field `NetworkPolicies` (where it belongs to)
- in field `DefaultNetworkPolicy`

The name of the default policy is not contained at all.

Replace field `DefaultNetworkPolicy` by `DefaultNetworkProfile`, which
contains the key of the default network policy to be found in
`NetworkProfiles`.

Consumers now have both the name of the default network profile as well
as all network profiles available.