Do not swallow errors, but panic.
We have cases where we know there cannot be en error when calling
changeStatusAndStoreForRetry. Nevertheless this function returns
an error (... always nil in the cases described above). In order
to be able to detect unexpected behaviour of the coding we panic in
such cases instead of simply silently swallowing these unexpected
errors.