[CLOUDCIFEAT1-130] Avoid non atomic updates (#248)
Up to now each and every change was "commited" directly. Now we apply several changes to the memory representation of a pipeline run and "commit" these changes together. With that there is no (short) period of time where the pipeline run is inconsistent.

There are also a lot of other changes more the less related (e.g. affecting error handling)

Co-authored-by: Matthias Rinck <m.rinck@sap.com>