Fix: Tekton fails to start JFR pod if container registry rate limit is exceeded
Steward uses Tekton to run JFR pods for pipeline runs. The respective
ClusterTask contains a pod template that does _not_ specify the
entrypoint of the JFR container. Therefore Tekton tries to obtain the
entrypoint by downloading the container image manifest from the
registry. This may fail if the registry's rate limit is exceeded. As a
workaround, specify the JFR container entrypoint in the ClusterTask.