Fix metric `steward_pipelineruns_ongoing_state_duration_periodic_observations_seconds` (#322)
The duration of an observed pipelinerun's current phase is wrongly
calculated based on the start of the pipeline run (`status.startedAt`)
instead of the start of the current phase
(`status.stateDetails.startedAt`).

Consequently no observations are made for phases `preparing` and
`waiting`, because there was no start time yet. For observations of
phases `cleaning` and `finished` the duration is too long.